git subrepo pull hc
subrepo:
  subdir:   "hc"
  merged:   "73ad5c4"
upstream:
  origin:   "git@github.com:henke96/hc.git"
  branch:   "master"
  commit:   "73ad5c4"
git-subrepo:
  version:  "0.4.6"
  origin:   "???"
  commit:   "???"

+++ hc/src/hc/wasm/heap.c
@@ -3,17 +3,17 @@
 #define heap_WASM_PAGE_SHIFT 16
 #define heap_WASM_PAGE_SIZE (1 << (heap_WASM_PAGE_SHIFT))
 
-static void *heap_reserve(int64_t size) {
-    uint32_t numPages = (uint32_t)(((uint64_t)size + heap_WASM_PAGE_SIZE - 1) >> heap_WASM_PAGE_SHIFT);
-    ssize_t mem = (ssize_t)hc_WASM_MEMORY_GROW(numPages);
+static void *heap_reserve(int32_t size) {
+    uint32_t numPages = ((uint32_t)size + heap_WASM_PAGE_SIZE - 1) >> heap_WASM_PAGE_SHIFT;
+    int32_t mem = (int32_t)hc_WASM_MEMORY_GROW(numPages);
     if (mem == -1) return NULL;
     return (void *)mem;
 }
 
-static void heap_unreserve(hc_UNUSED void *address, hc_UNUSED int64_t size) {}
+static void heap_unreserve(hc_UNUSED void *address, hc_UNUSED int32_t size) {}
 
-static int32_t heap_commit(hc_UNUSED void *address, hc_UNUSED int64_t size) {
+static int32_t heap_commit(hc_UNUSED void *address, hc_UNUSED int32_t size) {
     return 0;
 }
 
-static void heap_decommit(hc_UNUSED void *address, hc_UNUSED int64_t size) {}
+static void heap_decommit(hc_UNUSED void *address, hc_UNUSED int32_t size) {}
